Leech_Jack_Final_Proj/requirements.txt
# requirements.txt
openai
langchain
faiss-cpu
streamlit
python-dotenv
PyPDF2
tiktoken
langchain>=0.1.0
langchain-community>=0.0.10
requests
